+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - implement validation for models
 - wrap `sqlalchemy.orm.relationship` with the configured `Query` class
+- override the `alembic` command to customize the generated migrations templates
 
 ## v0.1.0 (2018/09/24)